feat: Implement universal data assimilation framework and examples
This commit introduces a comprehensive data assimilation framework to the CHS SDK, designed to support online adaptive identification and assimilation for various model types.

The key changes include:

1.  **Core Architecture Refactoring**:
    *   A new abstract base class, `BaseDataAssimilation`, is created in `control/data_assimilation.py` to define a common interface for all filtering algorithms.
    *   The existing `KalmanFilter` has been refactored to inherit from this new base class, aligning it with the new framework.

2.  **Advanced Filtering Algorithms**:
    *   `ExtendedKalmanFilter` (EKF) has been implemented to handle non-linear systems through Jacobian-based linearization.
    *   `EnsembleKalmanFilter` (EnKF) has been implemented to support high-dimensional, non-linear systems and state-parameter estimation using an ensemble approach.

3.  **Comprehensive Example Library**:
    *   `case_18_ekf_st_venant_state.py`: Demonstrates using EKF for real-time state correction of a 1D St. Venant hydrodynamic model.
    *   `case_19_kf_dynamic_tracking_transient.py`: Shows how a standard KF can help a simple `MuskingumChannelModel` track the output of a complex `StVenantModel`.
    *   `case_20_enkf_hydrology_parameter.py`: Illustrates using EnKF for online estimation of a time-varying parameter (SCS Curve Number) within a semi-distributed hydrological model.

4.  **Supporting Model Enhancements**:
    *   Necessary methods (`set_state`, `get_node`, etc.) and bug fixes (dtype issues, Kalman gain calculation) were implemented in the `StVenantModel`, `HydrodynamicNetwork`, and `Solver` classes to support the integration with the new assimilation framework.

This new framework provides a powerful and extensible foundation for advanced online calibration, real-time correction, and forecasting capabilities within the CHS platform.

+++ water_system_sdk/src/water_system_simulator/hydrodynamics/solver.py
@@ -40,9 +40,9 @@ class Solver:
         self.num_vars = self.num_nodes + self.num_reaches + self.num_structures
 
     def solve_step(self, dt: float):
-        H_old = np.array([node.head for node in self.network.nodes])
-        Q_old = np.array([reach.discharge for reach in self.network.reaches])
-        S_old = np.array([struct.discharge for struct in self.network.structures])
+        H_old = np.array([node.head for node in self.network.nodes], dtype=float)
+        Q_old = np.array([reach.discharge for reach in self.network.reaches], dtype=float)
+        S_old = np.array([struct.discharge for struct in self.network.structures], dtype=float)
 
         H_new = np.copy(H_old)
         Q_new = np.copy(Q_old)
